+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - physical-storage-api: allow passing pv/vg/lv/zpool create arguments
 - Encryption: Added systemd-creds support for masterpassphrase.
+- DRBD: Added Linstor/Drbd/auto-block-size
 
 ### Changed